Rework RunLengthEncoder.
The existing RunLengthEncoder relies on being able to read ahead
a few bytes, and then decide that it wants to ask the caller for
more data.

Unfortunately, this means that the clist cmd_compress_bitmap routine
can only call it in the "here is all the data in a solid block" case,
not in the "here is the data a line at a time" case.

This will become even more of a limitation when I rework
cmd_compress_bitmap to avoid overrun reads in a future commit.

The primary difference here is that we never backtrack on our reads
within the compression routine. We keep the last 1, 2 or 3 bytes
read in the state as n0, n1 and n2, and we insert literal bytes into a
array within the state for copying out later.

Because we now allow cmd_compress_bitmap to run in cases when it
didn't before, we trip over a bug in there. If height > 0 and
raster < width_bytes (for instance when raster = 0, so we can repeat
the same line several times), on 64bit builds the pointer
arithmetic goes wrong, and we end up accessing illegal memory.
Fix with a simple cast to int.

+++ base/gxclbits.c
@@ -94,7 +94,7 @@ cmd_compress_bitmap(stream_state * st, const byte * data, uint width_bits,
                 status = -1;
                 break;
             }
-            r.ptr += raster - width_bytes;
+            r.ptr += (int)(raster - width_bytes);
         }
         if (status == 0)
             status = (*st->templat->process) (st, &r, pw, true);
